Add generation for inputs and password input component
