+++ TeachAndTest-client/src/app/shared/components/controls/password-input/password-input.component.html
@@ -1,1 +1,55 @@
-<p>password-input works!</p>
+<div class="password-input">
+  <div class="password-input__body">
+    <div class="password-input__content">
+      <label
+        [for]="name"
+        class="password-input__label"
+        [ngClass]="{
+          'password-input__label--upper': i.value
+        }"
+      >
+        {{ label }}
+      </label>
+      <input
+        #i
+        (input)="inputHandler($event)"
+        (blur)="onTouched()"
+        [disabled]="disabled"
+        [name]="name"
+        [type]="type"
+        [autocomplete]="autocomplete"
+        class="password-input__input"
+        [ngClass]="{
+          'password-input__input--invalid':
+            (controlDir.control.touched ||
+              activated) &&
+            controlDir.control.invalid,
+          'password-input__input--valid':
+            controlDir.control.valid
+        }"
+      />
+
+      <a
+         class="password-control"
+        (click)="changeType($event)"
+        [ngClass]="{
+          'view': type == 'text'
+        }"
+      ></a>
+    </div>
+    <button>change visibility</button>
+    <div
+      *ngIf="
+        (controlDir.control.touched || activated) &&
+        controlDir.control.invalid
+      "
+      class="password-input__error-msg"
+    >
+      <div *ngFor="let error of getErrors()">
+        {{ error }}
+      </div>
+
+      {{ errorMessage }}
+    </div>
+  </div>
+</div>
