+++ TeachAndTest-client/src/app/shared/components/controls/password-input/password-input.component.scss
@@ -1,0 +1,55 @@
+.password-input {
+	&__body {
+    padding: 4px;
+  }
+  &__content{
+    position: relative;
+    font-size: 3rem;
+    width: fit-content;
+  }
+	&__label {
+    display: block;
+    position: relative;
+    letter-spacing: 1px;
+    z-index: -1;
+    bottom: -3.5rem;
+    transition: linear .15s;
+    &--upper{
+      bottom: 0;
+    }
+  }
+
+	&__input {
+    font-size: inherit;
+    opacity: 0.7;
+    border: 0;
+    border-bottom: 3px solid;
+    border-radius: 2px;
+    outline: none;
+    &--invalid{
+      border-bottom-color: lightsalmon;
+      background-color: lightcoral;
+    }
+    &--valid{
+      border-bottom-color: limegreen;
+    }
+   }
+
+   &__error-msg{
+      font-size: 2rem;
+      color: red;
+   }
+
+  .password-control {
+    position: absolute;
+    right: 10px;
+    bottom: 10px;
+    display: inline-block;
+    width: 20px;
+    height: 20px;
+    background: url(https://snipp.ru/demo/495/view.svg) 0 0 no-repeat;
+  }
+  .password-control.view {
+    background: url(https://snipp.ru/demo/495/no-view.svg) 0 0 no-repeat;
+  }
+}
