+++ TeachAndTest-client/src/app/shared/components/controls/password-input/password-input.component.ts
@@ -5,17 +5,29 @@ import {
   Self,
 } from '@angular/core';
 import { NgControl } from '@angular/forms';
-import { ReactiveInputComponent } from '@shared/components/controls/reactive-input/reactive-input.component';
+import { CustomInputComponent } from '@shared/components/controls/custom-input/custom-input.component';
 
 @Component({
   selector: 'app-password-input',
   templateUrl: './password-input.component.html',
   styleUrls: ['./password-input.component.scss'],
 })
-export class PasswordInputComponent {
+export class PasswordInputComponent extends CustomInputComponent {
   constructor(public controlDir: NgControl) {
-    //super(controlDir);
+    super(controlDir);
   }
 
-  ngOnInit(): void {}
+  changeType(event: Event){
+    event.preventDefault();
+    if(this.type == "text"){
+      this.type = "password"
+    }else{
+      this.type = "text"
+    }
+  }
+
+  ngOnInit(){
+    this.type = "password";
+    super.ngOnInit();
+  }
 }
