+++ TeachAndTest-client/src/app/test/test.component.html
@@ -5,14 +5,13 @@
   ></app-reactive-input>
   {{ reactive.valid }}
 </form>
-{{fg.valid}}
+{{ fg.valid }}
 <hr />
 <app-reactive-input
   [(ngModel)]="model"
   pattern="^1.*"
 ></app-reactive-input>
 
-<app-password-input>
-
+<app-password-input [formControl]="this.password">
 </app-password-input>
 {{ model }}
